Created test script for case start SDL with Preloaded PT file without read permissions
